+++ src/app/pages/day/day.page.scss
