+++ src/app/pages/terms/terms-routing.module.ts
