+++ src/app/pages/terms/terms.page.html
