+++ src/app/pages/terms/terms.page.spec.ts
