+++ src/app/pages/terms/terms.page.ts
