--- src/app/pipes/index.ts
@@ -1,8 +1,0 @@
-import { NgModule } from "@angular/core";
-@NgModule({
-  imports: [],
-  exports: [],
-  declarations: [],
-  providers: [],
-})
-export class AppPipesModule {}
